Bugfixes + highlighting errors
Technically, trailing whitespace is not an error, but it has to be
clearly visible to the user to prevent mistakes.

Also, this syntax file has been a huge trial-and-error. I know RegEx,
but I'm not familiar with best practices regarding Vim sytax files.
(Insert meme picture of the dog "I have no idea what I'm doing".)

Still, this syntax seems to work for my cases.